+++ scripts/generate-client.sh
@@ -9,7 +9,7 @@ set -a
 set +a
 
 cd backend
-source venv/bin/activate
+source .venv/bin/activate
 python -c "import app.main; import json; print(json.dumps(app.main.app.openapi()))" > ../openapi.json
 cd ..
 mv openapi.json frontend/